### 2025-01-27 #7732 @editorialbot
[PRE REVIEW]: GrIML: A Python package for investigating Greenland’s ice-marginal lakes under a changing climate
**Submitting author:** @PennyHow (<a href="http://orcid.org/0000-0002-8088-8497">Penelope How</a>)
**Repository:** https://github.com/GEUS-Glaciology-and-Climate/GrIML
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
