+++ paper/paper.md
@@ -89,7 +89,7 @@ Full installation instructions are provided at [https://griml.readthedocs.io/en/
 
 ## Tutorials and guides
 
-Two sets of tutorials are provided with `GrIML` for 1) using the processing pipeline ([https://griml.readthedocs.io/en/latest/tutorials-processing.html](https://griml.readthedocs.io/en/latest/tutorials-processing.html)); and 2) handling the produced dataset ([https://griml.readthedocs.io/en/latest/tutorials-data.html](https://griml.readthedocs.io/en/latest/tutorials-data.html)). The processing tutorials guide the user through each processing step from the initial raster classifications to a fully compiled ice-marginal lake dataset. The dataset tutorials provide an overview for how to handle and analyse an ice-marginal lake inventory produced with `GrIML`. This includes examples using the ice-marginal lakes dataset on the GEUS Dataverse [@How2025a], primarily handled with Geopandas [@Kelsey2020] to create spatial plots and time-series (\autoref{fig:tutorial}).
+Two sets of tutorials are provided with `GrIML` for 1) using the processing pipeline ([https://griml.readthedocs.io/en/latest/tutorials-processing.html](https://griml.readthedocs.io/en/latest/tutorials-processing.html)); and 2) handling the produced dataset ([https://griml.readthedocs.io/en/latest/tutorials-data.html](https://griml.readthedocs.io/en/latest/tutorials-data.html)). The processing tutorials guide the user through each processing step from the initial raster classifications to a fully compiled ice-marginal lake dataset. The dataset tutorials provide an overview for how to handle and analyse an ice-marginal lake inventory produced with `GrIML`. This includes examples using the ice-marginal lakes dataset on the GEUS Dataverse [@How2025a], primarily handled with Geopandas [@Geopandas2024] to create spatial plots and time-series (\autoref{fig:tutorial}).
 
 ![An example plot output from the dataset tutorial, where the change in the number of lakes across the ice-marginal lake inventory series is visualised. Each bar represents the number of lakes in one inventory year, with each coloured stage indicating the number of lakes for a given region. \label{fig:tutorial}](https://github.com/GEUS-Glaciology-and-Climate/GrIML/blob/main/docs/figures/iml_time_series_plot.png?raw=true)
 
