+++ paper/paper.bib
@@ -44,7 +44,7 @@ journal = {Geophysical Research Letters},
 volume = {49},
 number = {12},
 pages = {e2022GL099276},
-doi = {https://doi.org/10.1029/2022GL099276},
+doi = {10.1029/2022GL099276},
 url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2022GL099276},
 year = {2022}
 }
@@ -84,7 +84,7 @@ pages = {18-27},
 year = {2017},
 note = {Big Remotely Sensed Data: tools, applications and experiences},
 issn = {0034-4257},
-doi = {https://doi.org/10.1016/j.rse.2017.06.031},
+doi = {10.1016/j.rse.2017.06.031},
 url = {https://www.sciencedirect.com/science/article/pii/S0034425717302900},
 author = {Noel Gorelick and Matt Hancher and Mike Dixon and Simon Ilyushchenko and David Thau and Rebecca Moore},
 keywords = {Cloud computing, Big data, Analysis, Platform, Data democratization, Earth Engine},
@@ -126,13 +126,13 @@ keywords = {Cloud computing, Big data, Analysis, Platform, Data democratization,
 	year = {Submitted},
 }
 
-@article{Kelsey2020,
-  author = {Kelsey, J. and others},
-  title = {geopandas/geopandas: v0.8.1},
-  year = {2020},
+@article{Geopandas2024,
+  author = {Van den Bossche, J. and others},
+  title = {geopandas/geopandas: v1.0.1},
+  year = {2024},
   journal = {Zenodo},
-  doi = {10.5281/zenodo.394676},
-  url = {https://doi.org/10.5281/zenodo.394676}
+  doi = {10.5281/zenodo.12625316},
+  url = {https://doi.org/10.5281/zenodo.12625316},
 }
 
 @article{Kjeldsen2017,
