+++ paper/paper.bib
@@ -1,11 +1,3 @@
-@article{AMAP2021,
-  title = {Arctic Climate Change Update 2021: Key Trends and Impacts. Summary for Policy-makers},
-  author = {{Arctic Monitoring and Assessment Programme (AMAP)}},
-  year = {2021},
-  journal = {Arctic Monitoring and Assessment Programme (AMAP), Tromsø, Norway},
-  note = {16 pp}
-}
-
 @article{Andreassen2022,
   author = {Andreassen, L. M. and Nagy, T. and Kjøllmoen, B. and Leigh, J. R.},
   title = {An inventory of Norway’s glaciers and ice-marginal lakes from 2018–19 Sentinel-2 data},
@@ -15,38 +7,6 @@
   pages = {1085--1106},
   year = {2022},
   doi = {10.1017/jog.2022.20}
-}
-
-@article{Carrivick2016,
-  author = {Carrivick, J. L. and Tweed, F. S.},
-  title = {A global assessment of the societal impacts of glacier outburst floods},
-  journal = {Global and Planetary Change},
-  volume = {144},
-  pages = {1--16},
-  year = {2016},
-  doi = {10.1016/j.gloplacha.2016.07.001}
-}
-
-@article{Carrivick2019,
-  author = {Carrivick, J. L. and Tweed, F. S.},
-  title = {A review of glacier outburst floods in Iceland and Greenland with a megafloods perspective},
-  journal = {Earth-Science Reviews},
-  volume = {196},
-  pages = {102876},
-  year = {2019},
-  doi = {10.1016/j.earscirev.2019.102876}
-}
-
-@article{Carrivick2022,
-author = {Carrivick, J. L. and How, P. and Lea, J. M. and Sutherland, J. L. and Grimes, M. and Tweed, F. S. and Cornford, S. and Quincey, D. J. and Mallalieu, J.},
-title = {Ice-Marginal Proglacial Lakes Across Greenland: Present Status and a Possible Future},
-journal = {Geophysical Research Letters},
-volume = {49},
-number = {12},
-pages = {e2022GL099276},
-doi = {10.1029/2022GL099276},
-url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2022GL099276},
-year = {2022}
 }
 
 @article{Chen2021,
@@ -90,16 +50,6 @@ author = {Noel Gorelick and Matt Hancher and Mike Dixon and Simon Ilyushchenko a
 keywords = {Cloud computing, Big data, Analysis, Platform, Data democratization, Earth Engine},
 }
 
-@article{Grinsted2017,
-  author = {Grinsted, A. and Hvidberg, C. S. and Campos, N. and Dahl-Jensen, D.},
-  title = {Periodic outburst floods from an ice-dammed lake in East Greenland},
-  journal = {Scientific Reports},
-  volume = {7},
-  pages = {9966},
-  year = {2017},
-  doi = {10.1038/s41598-017-07960-9}
-}
-
 @article{How2021,
   author = {How, P. and Messerli, A. and M{\"a}tzler, E. and Santoro, M. and Wiesmann, A. and Caduff, R. and Langley, K. and Bojesen, M. H. and Paul, F. and K{\"a}{\"a}b, A. and Carrivick, J. L.},
   title = {Greenland-wide inventory of ice marginal lakes using a multi-method approach},
@@ -133,17 +83,6 @@ keywords = {Cloud computing, Big data, Analysis, Platform, Data democratization,
   journal = {Zenodo},
   doi = {10.5281/zenodo.12625316},
   url = {https://doi.org/10.5281/zenodo.12625316},
-}
-
-@article{Kjeldsen2017,
[…]
