### 2025-01-28 c5f8006
JOSS pre review edits (#21)
* dois updated

* word count reduced
