+++ paper/paper.md
@@ -39,14 +39,9 @@ Initial rasterised binary classifications denoting water bodies can be inputted 
 3. Showcase a transparent workflow that, in turn, produces an open and reproducible ice-marginal lake dataset that adheres to the FAIR principles [@Wilkinson2016]
 4. Produce inventories that map the areal extent and abundance of ice-marginal lakes across Greenland, which demonstrate ice-marginal lake evolution under a changing climate
 
-There have been many different approaches to classifying ice-marginal lakes with remote sensing techniques [@How2021;@Chen2021;@Andreassen2022;@Rick2022;@Wieczorek2023;@Domgaard2024]. Packages exist for handling satellite and spatial data, such as GrIML's two key dependencies, Geopandas [@Kelsey2020] and Rasterio [@Gillies2019], as well as others such as SentinelHub [@Sentinelhub2024] and Google Earth Engine [@Gorelick2017]. Remote sensing object classification and post-processing routines are usually available in connection with scientific publications, however, few are available as open, deployable packages. The `GrIML` Python package addresses this gap, for the benefit of the future generation of ice-marginal lake inventories and for others in the scientific community to adapt and use themselves. For example, it is intended to continue using the `GrIML` package for the production of future additions to the Greenland ice-marginal lake inventory series; in which case a standardised and version-controlled processing pipeline is essential [@How2025a].
+There have been many different approaches to classifying ice-marginal lakes with remote sensing techniques [@How2021;@Chen2021;@Andreassen2022;@Rick2022;@Wieczorek2023;@Domgaard2024]. Packages exist for handling satellite and spatial data, such as GrIML's two key dependencies, Geopandas [@Kelsey2020] and Rasterio [@Gillies2019], as well as others such as SentinelHub [@Sentinelhub2024] and Google Earth Engine [@Gorelick2017]. Remote sensing object classification and post-processing routines are usually available in connection with scientific publications, however, few are available as open, deployable packages. The `GrIML` Python package addresses this gap, for the benefit of the future generation of ice-marginal lake inventories and for others in the scientific community to adapt and use themselves. 
 
-
-## Scientific background and context
-
-The Greenland Ice Sheet and its peripheral glaciers and ice caps (PGICs) are anticipated to be the largest contributors to sea level rise during the 21st century [@AMAP2021]. While projections typically assume meltwater flows directly to the ocean, a significant portion is temporarily stored in ice-marginal lakes along the ice sheet's edges and near PGICs. These lakes form where meltwater is trapped at glacier termini, often exhibiting dynamic behaviours such as sudden drainage events that lead to Glacial Lake Outburst Floods (GLOFs) [@Carrivick2019]. Such events can affect glacier dynamics, sedimentation, and downstream ecosystems [@Grinsted2017;@Kjeldsen2017]. Ice-marginal lakes also drive thermo-mechanical processes, including lacustrine melting and calving, which accelerate ice mass loss [@Sutherland2020]. As glaciers retreat under warming conditions, these lakes are expected to increase in size and prevalence, amplifying proglacial melt rates and GLOF occurrences [@Carrivick2016;@Carrivick2022].
-
[…]
