### 2025-01-28 #7732 @PennyHow
Hi @kthyng, I've just edited the DOIs and reduced the length of the paper (as I saw the word count was a little over the recommendation). I have checked and think everything else fits the requirements for JOSS, but please let me know if there are any issues you would like me to address. As for reviewers, I have seen 5 reviewers that I think would be suitable.

ArcticSnow
rwegener2
trchudley
laurenguido
facusapienza21
