### 2025-02-10 #7732 @kthyng
@PennyHow:
* Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "norway" isn't capitalized but I think there were more.
* Your Geopandas inline ref was a "?"
* We have a backlog of submissions so I will add this to our waitlist. Thanks for your patience.
