### 2025-02-16 0756740
error estimate calculations added
