+++ paper/paper.bib
@@ -117,7 +117,7 @@ keywords = {Cloud computing, Big data, Analysis, Platform, Data democratization,
 	language = {English},
 	number = {v1.0.0},
 	journal = {Sinergise Solutions d.o.o., a Planet Labs company},
-	author = {Sentinel Hub},
+	author = {{Sentinel Hub}},
 	year = {2024},
 }
 
