### 2025-02-16 #7732 @PennyHow
> Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "norway" isn't capitalized but I think there were more.

Thanks for pointing this out. The references have now been changed so that countries and other pronouns are capitalized

> Your Geopandas inline ref was a "?"

This has now been updates with the correct reference

> We have a backlog of submissions so I will add this to our waitlist. Thanks for your patience.

Thanks for notifying me about this, and please let me know when you are ready to proceed. In the meantime, I will regenerate the pdf and checks.
