### 2025-02-16 67e4545
Bugfix/joss edits (#22)
* dois updated

* word count reduced

* error estimate calculations added

* references capitalised

* typo
