+++ paper/paper.md
@@ -39,7 +39,7 @@ Initial rasterised binary classifications denoting water bodies can be inputted 
 3. Showcase a transparent workflow that, in turn, produces an open and reproducible ice-marginal lake dataset that adheres to the FAIR principles [@Wilkinson2016]
 4. Produce inventories that map the areal extent and abundance of ice-marginal lakes across Greenland, which demonstrate ice-marginal lake evolution under a changing climate
 
-There have been many different approaches to classifying ice-marginal lakes with remote sensing techniques [@How2021;@Chen2021;@Andreassen2022;@Rick2022;@Wieczorek2023;@Domgaard2024]. Packages exist for handling satellite and spatial data, such as GrIML's two key dependencies, Geopandas [@Kelsey2020] and Rasterio [@Gillies2019], as well as others such as SentinelHub [@Sentinelhub2024] and Google Earth Engine [@Gorelick2017]. Remote sensing object classification and post-processing routines are usually available in connection with scientific publications, however, few are available as open, deployable packages. The `GrIML` Python package addresses this gap, for the benefit of the future generation of ice-marginal lake inventories and for others in the scientific community to adapt and use themselves. 
+There have been many different approaches to classifying ice-marginal lakes with remote sensing techniques [@How2021;@Chen2021;@Andreassen2022;@Rick2022;@Wieczorek2023;@Domgaard2024]. Packages exist for handling satellite and spatial data, such as GrIML's two key dependencies, Geopandas [@Geopandas2024] and Rasterio [@Gillies2019], as well as others such as SentinelHub [@Sentinelhub2024] and Google Earth Engine [@Gorelick2017]. Remote sensing object classification and post-processing routines are usually available in connection with scientific publications, however, few are available as open, deployable packages. The `GrIML` Python package addresses this gap, for the benefit of the future generation of ice-marginal lake inventories and for others in the scientific community to adapt and use themselves. 
 
 The 2016-2023 Greenland ice-marginal lake inventory series was produced with the `GrIML` Python package to address knowledge gaps regarding sea level contribution, and provide an overview of regions where lacustrine ice loss and processes are prominent [@How2025a;@How2025b]. The dataset consists of a series of annual inventories, mapping the presence and extent of ice-contact lakes across Greenland (\autoref{fig:inventory}). The annual inventory series spans the entirety of Greenland, including all terrestrial regions. Thus far, there are 8 annual inventories, covering 2016 to 2023, where one inventory represents one year. It is intended to continue using the `GrIML` package for the production of future additions to the Greenland ice-marginal lake inventory series; in which case a standardised and version-controlled processing pipeline is essential [@How2025a].
 
