+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{Andreassen2022,
   author = {Andreassen, L. M. and Nagy, T. and Kjøllmoen, B. and Leigh, J. R.},
-  title = {An inventory of Norway’s glaciers and ice-marginal lakes from 2018–19 Sentinel-2 data},
+  title = {An inventory of {Norway’s} glaciers and ice-marginal lakes from 2018–19 {Sentinel-2} data},
   journal = {Journal of Glaciology},
   volume = {68},
   number = {272},
@@ -11,7 +11,7 @@
 
 @article{Chen2021,
   author = {Chen, F. and Zhang, M. and Guo, H. and Allen, S. and Kargel, J. S. and Haritashya, U. K. and Watson, C.S.},
-  title = {Annual 30 m dataset for glacial lakes in High Mountain Asia from 2008 to 2017},
+  title = {Annual 30 m dataset for glacial lakes in {High} {Mountain} {Asia} from 2008 to 2017},
   journal = {Earth System Science Data},
   volume = {13},
   pages = {741--766},
@@ -21,7 +21,7 @@
 
 @article{Domgaard2024,
   author = {D{\o}mgaard, M. and Kjeldsen, K. and How, P. and Bj{\o}rk, A.},
-  title = {Altimetry-based ice-marginal lake water level changes in Greenland},
+  title = {Altimetry-based ice-marginal lake water level changes in {Greenland}},
   journal = {Communications Earth and Environment},
   volume = {5},
   pages = {365},
@@ -37,7 +37,7 @@
   url = "https://github.com/rasterio/rasterio"
 }
 @article{Gorelick2017,
-title = {Google Earth Engine: Planetary-scale geospatial analysis for everyone},
+title = {{Google} {Earth} {Engine}: Planetary-scale geospatial analysis for everyone},
 journal = {Remote Sensing of Environment},
 volume = {202},
 pages = {18-27},
@@ -62,7 +62,7 @@ keywords = {Cloud computing, Big data, Analysis, Platform, Data democratization,
 
 @article{How2025a,
   author = {How, P. and Petersen, D. and Karlsson, N. B. and Kjeldsen, K. K. and Raundrup, K. and Messerli, A. and Rutishauser, A. and Carrivick, J. L. and Lea, J. M. and Fausto, R. S. and Ahlstrøm, A. P. and Andersen, S. B.},
-  title = {Greenland Ice Marginal Lake Inventory annual time-series Edition 1 [dataset]},
+  title = {{Greenland} Ice Marginal Lake Inventory annual time-series {Edition} 1 [dataset]},
   year = {2025},
   journal = {GEUS Dataverse},
   doi = {10.22008/FK2/MBKW9N},
@@ -70,7 +70,7 @@ keywords = {Cloud computing, Big data, Analysis, Platform, Data democratization,
 }
 
 @article{How2025b,
-	title = {The Greenland Ice Marginal Lake Inventory Series from 2016 to 2023},
+	title = {The {Greenland} Ice Marginal Lake Inventory Series from 2016 to 2023},
 	journal = {Earth System Science Data},
 	author = {How, P. and Petersen, D. and Kjeldsen, K. K. and Raundrup, K. and Karlsson, N. B. and Messerli, A. and Rutishauser, A. and Carrivick, J. L. and Lea, J. M. and Fausto, R. S. and Ahlstrøm, A. P. and Andersen, S. B.},
 	year = {Submitted},
@@ -95,7 +95,7 @@ keywords = {Cloud computing, Big data, Analysis, Platform, Data democratization,
 
 @article{Mouginot2019,
   author = {Mouginot, J. and Rignot, E.},
-  title = {Glacier catchments/basins for the Greenland Ice Sheet},
[…]
