### 2025-03-03 #7732 @AdamRJensen
I'm serving as an editor at JOSS because I think open science matters, and JOSS does a pretty great job at pushing that agenda!
