### 2025-03-03 5ee7989
Editorial corrections for JOSS paper
