+++ paper/paper.md
@@ -23,16 +23,16 @@ bibliography: paper.bib
 
 # Summary
 
-The `GrIML` Python package is for processing classified water bodies from satellite imagery, and compiling classifications into a standardised inventory. It has been used to produce the Greenland ice-marginal lake inventory series, which maps the presence and extents of water bodies across Greenland that share a margin with the Greenland Ice Sheet and/or the surrounding ice caps and periphery glaciers since 2016 [@How2025a]. 
+The `GrIML` Python package is for processing classified water bodies from satellite imagery, and compiling classifications into a standardised inventory. It has been used to produce the Greenland ice-marginal lake inventory series, which maps the presence and extent of water bodies across Greenland that share a margin with the Greenland Ice Sheet and/or the surrounding ice caps and periphery glaciers since 2016 [@How2025a]. 
 
 ![An overview of the GrIML Python package workflow for generating ice-marginal lake inventories. The boxes refer to the four main modules and processing stages (reading from left to right), the coloured circles represent intermediary outputs to each processing stage, and the grey circles represent additional inputs. \label{fig:workflow}](https://github.com/GEUS-Glaciology-and-Climate/GrIML/blob/main/docs/figures/griml_workflow_without_gee.png?raw=true){ width=95% height=95% }
 
 
-Initial rasterised binary classifications denoting water bodies can be inputted to `GrIML` to convert, filter and merge into a unified ice-marginal lake vector dataset, populated with useful metadata and analysed with relevant statistical information (\autoref{fig:workflow}).
+Initial rasterised binary classifications denoting water bodies can be inputted to `GrIML` to convert, filter, and merge into a unified ice-marginal lake vector dataset, populated with useful metadata and analysed with relevant statistical information (\autoref{fig:workflow}).
 
 # Statement of need
 
-`GrIML` meets four main needs to users in the remote sensing and cryospheric science communities:
+`GrIML` meets four main needs of users in the remote sensing and cryospheric science communities:
 
 1. Provide a usable workflow for processing rasterised water body classifications
 2. Document the criteria for classifying an ice-marginal lake
@@ -41,9 +41,9 @@ Initial rasterised binary classifications denoting water bodies can be inputted 
 
 There have been many different approaches to classifying ice-marginal lakes with remote sensing techniques [@How2021;@Chen2021;@Andreassen2022;@Rick2022;@Wieczorek2023;@Domgaard2024]. Packages exist for handling satellite and spatial data, such as GrIML's two key dependencies, Geopandas [@Geopandas2024] and Rasterio [@Gillies2019], as well as others such as SentinelHub [@Sentinelhub2024] and Google Earth Engine [@Gorelick2017]. Remote sensing object classification and post-processing routines are usually available in connection with scientific publications, however, few are available as open, deployable packages. The `GrIML` Python package addresses this gap, for the benefit of the future generation of ice-marginal lake inventories and for others in the scientific community to adapt and use themselves. 
 
[…]
