### 2025-03-03 #7732 @AdamRJensen
@PennyHow I suggest including a grant number for the PROMICE project in the acknowledgements. Also, the developer installation instructions seem too specific to go in the paper but belong in the package documentation.
