### 2025-03-04 cf1bee8
Developer installation instructions removed
