### 2025-03-04 #7732 @AdamRJensen
> Hi [@AdamRJensen](https://github.com/AdamRJensen), thanks for your feedback and getting the ball rolling with this. I have addressed the issues you posted to the repo and looked at your PR - thank you! I also just removed the developer installation instructions from the paper, like you suggested.
>
> With regards to the grant number for PROMICE, I'm chasing this up with the PROMICE programme manager but, in general, PROMICE publications do not use the grant number. I suspect this is because PROMICE is a long-term project that has come from many different grant numbers. Please feel free to continue the pre-review steps whilst I see if I can find a solution to this.

@PennyHow That seems reasonable to me, please forget about the grant number. Go do some research instead! 😄
