### 2025-03-04 d66ffe1
Exception removed (#31)
