### 2025-03-05 3ee9f2b
Citations and DOIs updated
