+++ CITATION.cff
@@ -13,11 +13,14 @@ authors:
     orcid: 'https://orcid.org/0000-0002-8088-8497'
 identifiers:
   - type: doi
-    value: 10.5281/zenodo.11395471
+    value: https://doi.org/10.5281/zenodo.14718898
     description: Zenodo DOI for releases
   - type: doi
     value: https://doi.org/10.22008/FK2/MBKW9N
     description: Corresponding dataset
+  - type: doi
+    value: https://doi.org/10.5194/essd-2025-18
+    description: Data description paper
 repository-code: 'https://github.com/GEUS-Glaciology-and-Climate/GrIML'
 url: 'https://griml.readthedocs.io/'
 keywords:
