### 2025-03-05 496162a
Acknowledgements edit (#32)
* Acknowledgements edit

* Citations updated

* italics formatting

* DOI added

* Citations and DOIs updated

* version bumped

* version downgraded to bugfix

* development status changed

* DOI added to description paper
