+++ paper/paper.bib
@@ -36,6 +36,7 @@
   year =      {2013--},
   url = "https://github.com/rasterio/rasterio"
 }
+
 @article{Gorelick2017,
 title = {{Google} {Earth} {Engine}: Planetary-scale geospatial analysis for everyone},
 journal = {Remote Sensing of Environment},
@@ -73,7 +74,9 @@ keywords = {Cloud computing, Big data, Analysis, Platform, Data democratization,
 	title = {The {Greenland} Ice Marginal Lake Inventory Series from 2016 to 2023},
 	journal = {Earth System Science Data},
 	author = {How, P. and Petersen, D. and Kjeldsen, K. K. and Raundrup, K. and Karlsson, N. B. and Messerli, A. and Rutishauser, A. and Carrivick, J. L. and Lea, J. M. and Fausto, R. S. and Ahlstrøm, A. P. and Andersen, S. B.},
-	year = {Submitted},
+	year = {In Review},
+	doi = {10.5194/essd-2025-18},
+	url = {https://doi.org/10.5194/essd-2025-18}
 }
 
 @article{Geopandas2024,
