### 2025-03-05 #7732 @PennyHow
The PROMICE programme manager got back to me with an answer: There is no grant number as we are funded through “Forskningsreserven” which is part of the annual Fiscal Law (Finansloven) for Denmark.

He asked if I could edit the acknowledgements a little (mainly to better reflect this funding situation), which I have just completed ([#32](https://github.com/GEUS-Glaciology-and-Climate/GrIML/pull/32)).

I have bumped the package version and made a new release to reflect these pre-review changes. Please could the version for reviewing be changed to `v1.0.1`, @AdamRJensen?
