### 2025-03-05 #7732 @AdamRJensen
> I have bumped the package version and made a new release to reflect these pre-review changes. Please could the version for reviewing be changed to `v1.0.1`, [@AdamRJensen](https://github.com/AdamRJensen)?

It is likely that the review will result in more changes, so the typical process is that you make a final release once all reviewers have completed their reviews. It is then this final version that will be referenced in the JOSS paper.
