### 2025-03-19 #7927 @editorialbot
[REVIEW]: GrIML: A Python package for investigating Greenland’s ice-marginal lakes under a changing climate
**Submitting author:** @PennyHow (<a href="http://orcid.org/0000-0002-8088-8497">Penelope How</a>)
**Repository:** https://github.com/GEUS-Glaciology-and-Climate/GrIML
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.2
