### 2025-04-04 #7927 @rwegener2
Hi @PennyHow! This looks like a really nice package. I'm still mid review, but I have one question so far:

How is the [CURATED-ESA-GRIML-IML-fv1.gpkg](https://dataverse.geus.dk/file.xhtml?persistentId=doi:10.22008/FK2/MBKW9N/URDLNE&version=1.0) dataset from the GEUS Dataverse different from the other files in that data archive? I don't see a description of the CURATED file in the GEUS description, the dataset readme, or the readthedocs Dataset page.
