### 2025-04-08 #7927 @PennyHow
Oh! You are absolutely right that there is no description of this anywhere. I'll add it to the dataset readme and readthedocs and link to a PR here.

Essentially, this curated dataset consists of all ice-marginal lakes that have been manually identified - it forms part of the validation of the automated classification approach, comparing the automated and manually classified lakes to assess lake abundance.
