### 2025-04-09 6efb8cd
Curated file info added to documentation (#34)
* Curated data file defined (#33)

* curated datafile description added

* reformated file syntax

* reformated <> characters
