### 2025-04-14 #7927 @rwegener2
I have a few more general questions to help me orient before diving into the details of the review:
* What is the scope of this tool? Does it work outside of Greenland? Outside the Arctic? Does it work on other types of features (ex. a binary classification of urban vs. vegetated pixels)?
* The binary raster could have come from any source, correct? You mention satellite imagery in the package description, but could someone run this same pipeline on a binary raster generated from airborne or drone-based data equally well?

Overall, you've got a beautifully documented pipeline for the Greenland lakes inventory dataset. I'm trying to understand what other ways that someone might use this tool, given that you have already produced and made easily available the Greenland lakes dataset.
