### 2025-04-23 #7927 @PennyHow
Hi @rwegener2, in response to your questions:

> What is the scope of this tool? Does it work outside of Greenland? Outside the Arctic? Does it work on other types of features (ex. a binary classification of urban vs. vegetated pixels)?

The toolbox can be applied to any glacial environment, meaning that it is not limited to Greenland and the Arctic. The key limitation is that the classification criteria is very specific to classifying ice-marginal lakes, so it would not be possible to classify other features unless they fall within the same criteria as an ice-marginal lake.

There is potential interest in someone using the toolbox to classify ice-marginal lakes across Iceland, for example, which is absolutely feasible and exactly what I envisaged as a use-case.

> The binary raster could have come from any source, correct? You mention satellite imagery in the documentation, but could someone run this same pipeline on a binary raster generated from airborne or drone-based data equally well?

Yes, the binary raster can come from any source - satellite image, airborne imagery or drone orthophoto - on a regular spatial grid.

Perhaps these two answers should form some update to the documentation to ensure that all users understand this? I'm happy to add it.

> Overall, you've got a beautifully documented pipeline for the Greenland lakes inventory dataset. I'm trying to understand what other ways that someone might use this tool, given that you have already produced and made easily available the Greenland lakes dataset.

Why thank you, that's really nice to hear. As well as others using the toolbox to create ice-marginal lake datasets in other regions, one of the main reasons that I have submitted the toolbox here is because I want to continue using it to add additional data on a yearly basis in the future - having a peer-reviewed toolbox felt like a necessary step to ensure this is possible.
