### 2025-05-07 #7927 @facusapienza21
Hi @PennyHow ! Great work, very enjoying reviewing the work. A few questions as I make progress in the review:

- The script for running the tests include multiple tests using the `unittest` suite, which are further executed in CI. My question is: are all these tests executed? I ran them in local and obtained the [same results than in CI](https://github.com/GEUS-Glaciology-and-Climate/GrIML/actions/runs/14360247665/job/40259587143), which seems to show 4 different tests that a priori I cannot match with the ~9 tests defined inside `TestGrIML`. I am not familiar on how this suite works, so please let me know if I am missing something. Furthermore, I notice there are two other files in the test folder. Are these tests? Please notice that `process_example.py` includes hardcoded paths!

- Curiosity about the scope of the tool: can GrIML be used for other regions (e.g. Antarctica) or to create inventories of other features in the polar regions (e.g., supra glacier lakes or cravasses)?

- This may seem a very simple question, but what do the arrows in the third figure in `paper.md` represent? Are these the dependencies between modules?
