+++ docs/acknowledgements.md
@@ -4,7 +4,7 @@
 
 The inventory series of ice marginal lakes in Greenland has been produced as part of the European Space Agency (ESA) Living Planet Fellowship project "Examining GReenland’s Ice Marginal Lakes under a changing climate (GrIML)", which is a follow-on effort to the 2017 inventory of ice marginal lakes created under the European Space Agency (ESA) Climate Change Initiative (CCI) in Option 6 of the Glaciers_cci project (4000109873/14/I-NB). 
 
-Upkeep and continuation of the inventory series is supported by [PROMICE](https://promice.org], funded by the [Geological Survey of Denmark and Greenland (GEUS)](https://www.geus.dk/) and the Danish Ministry of Climate, Energy and Utilities under the Danish Cooperation for Environment in the Arctic (DANCEA), conducted in collaboration with DTU Space (Technical University of Denmark) and Asiaq Greenland Survey.
+Upkeep and continuation of the inventory series is supported by [PROMICE](https://promice.org], funded by the [Geological Survey of Denmark and Greenland (GEUS)](https://www.geus.dk/) and the Danish Ministry of Climate, Energy and Utilities, conducted in collaboration with DTU Space (Technical University of Denmark) and Asiaq Greenland Survey.
 
 
 ## Relevant links
