+++ docs/README-ESA-GRIML-IML-fv1.md
@@ -1,0 +1,96 @@
+# Greenland Ice-Marginal Lake Inventory annual time-series Edition 1
+
+Data available at [GEUS Dataverse](https://doi.org/10.22008/FK2/MBKW9N).
+
+Contact [pho@geus.dk](mailto:pho@geus.dk).
+
+## Table of contents
+
+- [Terms of use](#terms-of-use)
+- [Dataset Contents](#dataset-contents)
+	+ [Data format](#data-format)
+	+ [Metadata](#metadata)
+- [Acknowledgements](#acknowledgements)
+- [Relevant links](#relevant-links)
+
+## Terms of use
+
+If the data are presented or used to support results of any kind, please include references to the applicable publications:
+
+- *How, P. et al. (2025) "Greenland Ice-Marginal Lake Inventory annual time-series Edition 1". GEUS Dataverse. [https://doi.org/10.22008/FK2/MBKW9N](https://doi.org/10.22008/FK2/MBKW9N)*
+- *How, P. et al. (2021) "Greenland-wide inventory of ice marginal lakes using a multi-method approach". Sci. Rep. 11, 4481. [https://doi.org/10.1038/s41598-021-83509-1](https://doi.org/10.1038/s41598-021-83509-1)*
+- *How, P. et al. "Greenland ice-marginal lake inventory series from 2016 to 2023". In Prep.*
+- *How, P. "A Python package for investigating Greenland's ice-marginal lakes under a changing climate". In Prep.*
+
+And include the following statement in the acknowledgements:
+
+*"Ice-marginal lake data provided by the European Space Agency (ESA), and the Programme for Monitoring of the Greenland Ice Sheet (PROMICE) at the Geological Survey of Denmark and Greenland (GEUS) ([https://doi.org/10.22008/FK2/MBKW9N](https://doi.org/10.22008/FK2/MBKW9N))."*
+
+## Dataset contents
+
+This ice-marginal lake dataset is a series of annual inventories, mapping the extent and presence of lakes across Greenland that share a margin with the Greenland Ice Sheet and/or the surrounding ice caps and periphery glaciers. Specifically, the following files are included in this dataset:
+
+- *\<YYYYMMDD\>-ESA-GRIML-IML-\<version\>.gpkg*: Ice-marginal lake inventory for a specific year, provided as polygon vector features
+- *CURATED-ESA-GRIML-IML-\<version\>.gpkg*: All identified ice-marginal lakes across the inventory series, including manually-classified lakes, and provided as point vector features
+- *README-ESA-GRIML-IML-\<version\>.gpkg*: This dataset readme file
+
+The annual inventories provide a comprehensive record of all identified ice-marginal lakes, which have been detected using three independent remote sensing techniques:
+
+- DEM sink detection using the ArcticDEM (mosaic version 3)
+- SAR backscatter classification from Sentinel-1 imagery
+- Multi-spectral indices classification from Sentinel-2 imagery
+
+All data were compiled and filtered in a semi-automated approach, using a modified version of the [MEaSUREs GIMP ice mask](https://nsidc.org/data/NSIDC-0714/versions/1) to clip the dataset to within 1 km of the ice margin. Each detected lake was then verified manually. The methodology is open-source and provided in the associated [Github repository](https://github.com/GEUS-Glaciology-and-Climate/GrIML) for full reproducibility.
+
+The inventory series was created to better understand the impact of ice-marginal lake change on the future sea level budget and the terrestrial and marine landscapes of Greenland, such as its ecosystems and human activities. The dataset is a complete inventory series of Greenland, with no absent data.
+
+### Data format
+
[…]
