### 2025-05-08 0e24dff
Readme documentation updated (#35)
* readme updated

* script updates

* conflict resolve
