### 2025-05-09 cf683d2
Arrow dependency description added
