### 2025-05-09 #7927 @PennyHow
Thanks for taking a look at the paper and the software. In answer to your questions:

> The script for running the tests include multiple tests using the unittest suite, which are further executed in CI. My question is: are all these tests executed? I ran them in local and obtained the [same results than in CI](https://github.com/GEUS-Glaciology-and-Climate/GrIML/actions/runs/14360247665/job/40259587143), which seems to show 4 different tests that a priori I cannot match with the ~9 tests defined inside TestGrIML. I am not familiar on how this suite works, so please let me know if I am missing something.

The `TestGrIML` object has 9 functions, however, 5 of these are functions for generating random data to feed to the 4 testing functions. This is why there are only 4 tests that appear.

> Furthermore, I notice there are two other files in the test folder. Are these tests? Please notice that process_example.py includes hardcoded paths!

Ah! These should not be in the test folder. They have now been moved in [this PR](https://github.com/GEUS-Glaciology-and-Climate/GrIML/pull/36). I have also cleaned up other scripts where I could see paths had been hardcoded.

> Curiosity about the scope of the tool: can GrIML be used for other regions (e.g. Antarctica) or to create inventories of other features in the polar regions (e.g., supra glacier lakes or cravasses)?

GrIML can be applied to other regions for detecting ice-marginal lakes, and this is one of the main use cases which I see the package being incredibly useful for. The package could be extended to other glacial water features, such as supraglacial lakes, as they have a similar classification criteria. However, I do not think non-water features (such as crevasses) would be appropriate for classifying using GrIML. Do you think I should add this to the documentation to make it clear what GrIML can be used for?

> This may seem a very simple question, but what do the arrows in the third figure in paper.md represent? Are these the dependencies between modules?

Yes, the arrows represent module dependencies. I have added information to the figure caption in [this PR](https://github.com/GEUS-Glaciology-and-Climate/GrIML/pull/37).

> I was able to install GrIML both using the package manager (pip install griml) and in directly from the repo (pip install .). However, I was not able to create the stand alone environment as it is suggested at the end of the [documentation](https://griml.readthedocs.io/en/latest/installation.html). When doing this, conda cannot figure out the dependencies using the current channels. Have you seen this before? My guess is that the list of dependencies is too strict (maybe generate them with a conda env export --from-history will reduce the list to the main packages with their versions?)

Thanks for making me aware of this. I can look into making a better, simpler environment. Let me get back to you shortly when I have done this - I'll start looking into this now.
