### 2025-05-09 #7927 @PennyHow
> I was able to install GrIML both using the package manager (pip install griml) and in directly from the repo (pip install .). However, I was not able to create the stand alone environment as it is suggested at the end of the [documentation](https://griml.readthedocs.io/en/latest/installation.html). When doing this, conda cannot figure out the dependencies using the current channels. Have you seen this before? My guess is that the list of dependencies is too strict (maybe generate them with a conda env export --from-history will reduce the list to the main packages with their versions?)

> Thanks for making me aware of this. I can look into making a better, simpler environment. Let me get back to you shortly when I have done this - I'll start looking into this now.

I have made a very relaxed environment file with few pinned versions of packages. I have tested this and it is working for me. Please feel free to test it and let me know what you think @facusapienza21: https://github.com/GEUS-Glaciology-and-Climate/GrIML/pull/38
