### 2025-05-09 #7927 @facusapienza21
Thank you @PennyHow for addressing this points! A few comments.

`Do you think I should add this to the documentation to make it clear what GrIML can be used for?`
I think it will be nice to add to the documentation, yes. One more suggestion: I will add very early in the documentation what `GrIM` stands for. I know this is mentioned in the docs and it is something people working in the field know, but maybe it will help to have it very early in the docs!

Great work simplifying the environment file!
