### 2025-05-09 #7927 @facusapienza21
Hi @PennyHow ! Here a few more comments.

***Automated tests.*** Thank you for clarifying about the tests. I think then it will be useful to have some idea of the test coverage currently produced by these tests. Do you think this can be easily done?

***Community guidelines.*** I will expand this section a bit more in the docs. I don’t think it is really necessary to provide further step by step of how to contribute to the project, but it will be very useful to include links to more extensive tutorials of how to do this for the interested reader. For reference, I find the [OGGM documentation](https://docs.oggm.org/en/stable/contributing.html) a great reference for this!

#### Suggestions and minor corrections

- May I suggest including the code in the processing tutorials in the form of a Jupyter notebook so it can be executed alongside with the tutorial? Furthermore, this can be easily deployed using Binder (or Colab) so the user can directly run the example as it goes through the tutorial.
- I will move the background section further above in the documentation.
- Notice that some figures are not visible in [Background](https://griml.readthedocs.io/en/stable/background.html).

Other than that, all still looks in my side! I will move forward with some more revisions next week and hopefully end the review next week :)
