+++ docs/contributing.md
@@ -26,6 +26,10 @@ You can work directly with GrIML’s development if you have a contribution, suc
 
 In order to contribute, you will need your own fork of the GrIML GitHub repository to work on the code. Go to the repo and choose the Fork option. This now creates a copy in your own GitHub space, which is connected to the upstream GrIML repository.
 
+```{note}
+See [here](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/working-with-forks/fork-a-repo) for more on forking Github repositories.
+```
+
 ### Creating a development branch
 
 From your forked space, make sure you have a Python Environment for running GrIML, as described in the [installation](https://griml.readthedocs.io/en/latest/installation.html). Then create and checkout a branch to make your developments on.
@@ -53,4 +57,4 @@ And then open a pull request. Make sure to include the following in your pull re
 2. Details of what these changes are
 3. Any limitations or further development needed
 
-Your pull request will be reviewed and hopefully accepted. Following this, you will be listed as a contributor to GrIML.
+Your pull request will be reviewed and hopefully accepted. Following this, you will be listed as a contributor to GrIML. For a more thorough guide on pull requests, see the [Github docs pull requests guide](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/about-pull-requests). There is also a great overview of how to contribute to repositories [here](https://docs.github.com/en/get-started/exploring-projects-on-github/contributing-to-a-project) which takes you through all the steps described.
