### 2025-05-13 f78a88e
links included as notes
