+++ docs/contributing.md
@@ -18,6 +18,10 @@ print(metadata.version('griml'))
 Before submitting an issue, please make sure that your installation is correct and working from either the pip installation or the main branch of the GrIML repository.
 ```
 
+```{note}
+See [here](https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/creating-an-issue) for more on creating issues in a Github repository.
+```
+
 ## Contributing to GrIML
 
 You can work directly with GrIML’s development if you have a contribution, such as a solution to an issue or a suggestion for an enhancment.
@@ -57,4 +61,12 @@ And then open a pull request. Make sure to include the following in your pull re
 2. Details of what these changes are
 3. Any limitations or further development needed
 
-Your pull request will be reviewed and hopefully accepted. Following this, you will be listed as a contributor to GrIML. For a more thorough guide on pull requests, see the [Github docs pull requests guide](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/about-pull-requests). There is also a great overview of how to contribute to repositories [here](https://docs.github.com/en/get-started/exploring-projects-on-github/contributing-to-a-project) which takes you through all the steps described.
+Your pull request will be reviewed and hopefully accepted. Following this, you will be listed as a contributor to GrIML. 
+
+```{note}
+For a more thorough guide on pull requests, see the [Github docs pull requests guide](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/about-pull-requests). 
+```
+
+```{note}
+There is also a great overview of how to contribute to repositories [here](https://docs.github.com/en/get-started/exploring-projects-on-github/contributing-to-a-project) which takes you through all the steps described.
+```
