### 2025-05-13 #7927 @PennyHow
> I think it will be nice to add to the documentation, yes. One more suggestion: I will add very early in the documentation what GrIM stands for. I know this is mentioned in the docs and it is something people working in the field know, but maybe it will help to have it very early in the docs!

I have added what GrIML stands for in brackets after the package is first introduced in the Github readme and readthedocs landing page. These changes are in [PR #39](https://github.com/GEUS-Glaciology-and-Climate/GrIML/pull/39).

> Automated tests. Thank you for clarifying about the tests. I think then it will be useful to have some idea of the test coverage currently produced by these tests. Do you think this can be easily done?

I have to make some changes to the unit testing structure in order for the coverage tests to be properly implemented, but I think this is a great idea to add and I would like to do this. I'll come back to you when I have made the updates.

> Community guidelines. I will expand this section a bit more in the docs. I don’t think it is really necessary to provide further step by step of how to contribute to the project, but it will be very useful to include links to more extensive tutorials of how to do this for the interested reader. For reference, I find the [OGGM documentation](https://docs.oggm.org/en/stable/contributing.html) a great reference for this!

I have now also included links to some more extensive tutorials like you have suggested (in [PR #39](https://github.com/GEUS-Glaciology-and-Climate/GrIML/pull/39)).

> May I suggest including the code in the processing tutorials in the form of a Jupyter notebook so it can be executed alongside with the tutorial? Furthermore, this can be easily deployed using Binder (or Colab) so the user can directly run the example as it goes through the tutorial.

I can definitely provide a Jupyter Notebook version of the processing tutorial which I will add in a later PR. I would like to also look into Binder and/or Colab also, so let me get back to you on this.

> I will move the background section further above in the documentation.

Done! The change is in [PR #39](https://github.com/GEUS-Glaciology-and-Climate/GrIML/pull/39).

> Notice that some figures are not visible in Background.

They should now be there (from [PR #39](https://github.com/GEUS-Glaciology-and-Climate/GrIML/pull/39)) - this happened because of some repo re-structuring.
