### 2025-05-13 fe83fbb
JOSS review: docs updates (#39)
* docs updates

* links added

* links included as notes
