+++ README.md
@@ -2,7 +2,7 @@
 
 [![PyPI version](https://badge.fury.io/py/griml.svg)](https://badge.fury.io/py/griml) [![DOI](https://zenodo.org/badge/444752900.svg)](https://zenodo.org/badge/latestdoi/444752900) [![Documentation Status](https://readthedocs.org/projects/griml/badge/?version=latest)](https://griml.readthedocs.io/en/latest/?badge=latest) [![Build Status](https://img.shields.io/endpoint.svg?url=https%3A%2F%2Factions-badge.atrox.dev%2FPennyHow%2FGrIML%2Fbadge%3Fref%3Dmain&style=flat)](https://actions-badge.atrox.dev/PennyHow/GrIML/goto?ref=main)
 
-The **GrIML** processing package for classifying water bodies from satellite imagery using a multi-sensor, multi-method remote sensing approach. This workflow is used for the production of the [Greenland ice-marginal lake inventory series](https://doi.org/10.22008/FK2/MBKW9N), as part of the [ESA GrIML project](https://eo4society.esa.int/projects/griml/). This repository also holds all project-related materials.
+The **GrIML** (Investigating Greenland's ice marginal lakes under a changing climate) processing package for classifying water bodies from satellite imagery using a multi-sensor, multi-method remote sensing approach. This workflow is used for the production of the [Greenland ice-marginal lake inventory series](https://doi.org/10.22008/FK2/MBKW9N), as part of the [ESA GrIML project](https://eo4society.esa.int/projects/griml/). This repository also holds all project-related materials.
 
 
 ## Installation
