### 2025-05-14 522ed7b
Features/environment file update (#38)
* yml file trimming

* Delete other/reporting/ice_marginal_lake_inventory_essd_paper.pdf

* stray hardcoded path removed
