+++ environment.yml
@@ -3,153 +3,13 @@ channels:
   - defaults
   - conda-forge
 dependencies:
-  - _libgcc_mutex=0.1=conda_forge
-  - _openmp_mutex=4.5=2_kmp_llvm
-  - affine=2.4.0=pyhd8ed1ab_0
-  - attrs=23.2.0=pyh71513ae_0
-  - blas=1.0=mkl
-  - blosc=1.21.3=h6a678d5_0
-  - boost-cpp=1.81.0=he95ae9e_0
-  - bottleneck=1.3.7=py311hf4808d0_0
-  - branca=0.7.2=pyhd8ed1ab_0
-  - brotli=1.0.9=h9c3ff4c_4
-  - brotli-python=1.0.9=py311h6a678d5_7
-  - bzip2=1.0.8=h5eee18b_5
-  - c-ares=1.19.1=h5eee18b_0
-  - ca-certificates=2024.2.2=hbcca054_0
-  - cairo=1.16.0=hb05425b_5
-  - certifi=2024.2.2=pyhd8ed1ab_0
-  - cfitsio=4.2.0=hd9d235c_0
-  - charset-normalizer=3.3.2=pyhd8ed1ab_0
-  - click=8.1.7=unix_pyh707e725_0
-  - click-plugins=1.1.1=py_0
-  - cligj=0.7.2=pyhd8ed1ab_1
-  - contourpy=1.2.0=py311hdb19cb5_0
-  - curl=8.5.0=hdbd6064_0
-  - cycler=0.12.1=pyhd8ed1ab_0
-  - expat=2.6.2=h6a678d5_0
-  - fiona=1.9.5=py311ha02d727_0
-  - folium=0.16.0=pyhd8ed1ab_0
-  - fontconfig=2.14.2=h14ed4e7_0
-  - fonttools=4.25.0=pyhd3eb1b0_0
-  - freetype=2.12.1=h267a509_2
-  - freexl=1.0.6=h166bdaf_1
-  - gdal=3.6.2=py311hadb6153_4
-  - geopandas=0.14.4=pyhd8ed1ab_0
-  - geopandas-base=0.14.4=pyha770c72_0
-  - geos=3.11.1=h27087fc_0
-  - geotiff=1.7.1=h7a142b4_6
-  - giflib=5.2.1=h5eee18b_3
-  - glib=2.69.1=he621ea3_2
-  - hdf4=4.2.15=h9772cbc_5
-  - hdf5=1.12.2=nompi_h4df4325_101
-  - icu=70.1=h27087fc_0
-  - idna=3.7=pyhd8ed1ab_0
-  - intel-openmp=2023.1.0=hdb19cb5_46306
-  - jinja2=3.1.3=pyhd8ed1ab_0
-  - joblib=1.4.0=pyhd8ed1ab_0
-  - jpeg=9e=h166bdaf_1
-  - json-c=0.16=hc379101_0
-  - kealib=1.5.0=ha7026e8_0
-  - kiwisolver=1.4.4=py311h6a678d5_0
-  - krb5=1.20.1=h143b758_1
-  - lcms2=2.12=h3be6417_0
-  - ld_impl_linux-64=2.38=h1181459_1
-  - lerc=4.0.0=h27087fc_0
-  - libaec=1.1.3=h59595ed_0
-  - libcurl=8.5.0=h251f7ec_0
-  - libdeflate=1.17=h5eee18b_1
-  - libedit=3.1.20230828=h5eee18b_0
-  - libev=4.33=h516909a_1
-  - libexpat=2.6.2=h59595ed_0
-  - libffi=3.4.4=h6a678d5_0
-  - libgcc-ng=13.2.0=hc881cc4_6
[…]
