### 2025-05-14 6a60bcb
Features/add tutorials (#40)
* test notebook added

* testing playground changed

* Dataset tutorial added

* griml version call updated
