### 2025-05-14 cc90502
Explicit assert calls added
