+++ .github/workflows/unit_test.yml
@@ -1,3 +1,4 @@
+name: 'Unit tests'
 on:
   pull_request:
     types: [opened, reopened, synchronize, edited]
