+++ .github/workflows/unit_test_coverage.yml
@@ -10,7 +10,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ['3.10','3.11','3.12']
+        python-version: ['3.12']
     steps:
       - name: Install Python
         uses: actions/setup-python@v5
