### 2025-05-14 e19d47c
Features/unit test coverage (#42)
* Explicit assert calls added

* Coverage action added

* actions renamed

* coverage environment limited
