### 2025-05-14 #7927 @PennyHow
> Automated tests. Thank you for clarifying about the tests. I think then it will be useful to have some idea of the test coverage currently produced by these tests. Do you think this can be easily done?

> I have to make some changes to the unit testing structure in order for the coverage tests to be properly implemented, but I think this is a great idea to add and I would like to do this. I'll come back to you when I have made the updates.

A coverage report is now provided when a pull request is triggered. You can see an example of one of these [here](https://github.com/GEUS-Glaciology-and-Climate/GrIML/pull/42). I can set a coverage % threshold to block pull requests when the coverage is not high enough, however, I have kept these fairly low for now - I think the main thing is that it is reporting a coverage which can be viewed and assessed at the same time as a pull request. Hope this is what you were thinking and you are happy with it!
