### 2025-05-21 #7927 @AdamRJensen
Great to see so much discussion, and several nice improvements!

@rwegener2, @facusapienza21 does the recent discussion allow you to tick off some of the boxes on your review checklist?
