### 2025-05-30 #7927 @facusapienza21
Thank you @PennyHow ! The new features look amazing!!!

A few comments on the Notebook and Binder:

- [ ] Notice that as it is, the notebook does not run in Binder since there is a missing `wget` dependency. You can include this in  your environment file or install it directly in the notebook with `pip install wget`.
- [ ] It took me some time to find the notebook, since it is not very intuitive that this will be inside `tests`. I will recommend to move it to another folder (e.g., something like Tutorials) or to make Binder to automatically lunch this specific notebook (You can do this by adding a specific "File to Open" in the Binder website.
- [ ] The code in the Notebook looks great and I am so happy I can run the tutorial in this format! I think this is a great feature. Just as a matter of style, I will recommend to add more headers and descriptions in the Notebook to make it even more appealing to the user who opens and plays with the notebook (this is very a side suggestion, please don't feel forced to make further changes).

![Image](https://github.com/user-attachments/assets/b3cd6c72-5ce3-44a9-b648-5c33404a0116)
