### 2025-06-02 b619074
Features/joss edits (#43)
* wget dependency added

* notebook tutorials moved
