+++ environment.yml
@@ -13,3 +13,4 @@ dependencies:
   - scipy
   - pip:
     - griml
+    - wget
