### 2025-06-02 #7927 @PennyHow
> Notice that as it is, the notebook does not run in Binder since there is a missing wget dependency. You can include this in your environment file or install it directly in the notebook with pip install wget.

I've added wget to the environment file, as part of [#43](https://github.com/GEUS-Glaciology-and-Climate/GrIML/pull/43). The binder links have been updated in [#44](https://github.com/GEUS-Glaciology-and-Climate/GrIML/pull/44).

> It took me some time to find the notebook, since it is not very intuitive that this will be inside tests. I will recommend to move it to another folder (e.g., something like Tutorials) or to make Binder to automatically lunch this specific notebook (You can do this by adding a specific "File to Open" in the Binder website.

The Jupyter Notebooks have been moved to a directory named "tutorials" ([#43](https://github.com/GEUS-Glaciology-and-Climate/GrIML/pull/43)) and a new binder link( [#44](https://github.com/GEUS-Glaciology-and-Climate/GrIML/pull/44)) now points to the dataset handling tutorial.

> The code in the Notebook looks great and I am so happy I can run the tutorial in this format! I think this is a great feature. Just as a matter of style, I will recommend to add more headers and descriptions in the Notebook to make it even more appealing to the user who opens and plays with the notebook (this is very a side suggestion, please don't feel forced to make further changes). This code is the same that is shown in the documentation, so I wonder if there is a way to make this page the same as the notebook, so there is no need to "copy" information.

Thanks for this feedback. I've not worked with Jupyter Notebook much before, so I did not know that you could add headers and descriptions. I've added these now.
