### 2025-06-04 #7927 @rwegener2
Sorry to the team for the long delay in correspondence. May was a particularly busy month at work so I've been really occupied, but that commitment is over and I'm fired up to finish this review.

@PennyHow thanks for your explanations and quick updates to the package from my April inquiries! I think these improvements and the suggestions from @facusapienza21 have really made the package more robust and user friendly. I think this package is very close to being ready. I just have a few last comments, mostly related to helping users understand what the package is for an how to get started:

### Zooming in the clearest package description
This might just be me, but I still find the package description in [the docs intro](https://griml.readthedocs.io/en/stable/index.html#griml) a little bit confusing. I actually find the first sentence of your JOSS paper to be clearer "The GrIML Python package is for processing classified water bodies from satellite imagery, and compiling classifications into a standardised inventory." As far as I understand the package, I think the most important points to convey in the first 1-2 paragraphs of the docs home page are:
- this is a package for classifying ice-marginal lakes
- the package starts with/needs as an input a binary raster image from satellite (or aerial) imagery
- the package will produce a shapefile containing the (point? polygon?) locations of ice marginal lakes that are a certain size and distance from the ice sheet
- a dataset for Greenland already exists that has been produced by this package. But this package can be used in other ice-marginal areas or could be used to recreate the Greenland dataset with new inputs

Something that could be helpful is a short set of bullets like:
"This documentation page guides users through: 1) running the lake classification code and 2) provides examples of working with the Greenland dataset produced from this package".

You could also move the "If the GrIML package or the Greenland ice marginal lake inventory series are presented..." into a separate paragraph (or even a subsection called "Citing GrIML" or "Acknowledging GrIML").

I'm sorry if this feels nit-picky. The package description is usually the first thing a user reads, and they use it to decide if they want to use the package or not, so I think it's worth being intentional!

### Clarifying input data
I am still a little bit lost about what kind of data someone needs to start using this package. The clearest path seems to be using the GEE script to process Sentinel-2. @PennyHow, as a package maintainer would you like to support people bringing other datasets to use in your package? You certainly don't need to, this package is good as it is. Depending on what you want to support, though, I would either:
1. [if people should only bring the Sentinel-2 processed imagery] --> shift the wording of the tutorial slightly to just say that the first processing step is to run that provided script on GEE (As opposed to "If you have followed...")
2. [if people could bring other data or rasters classified from other sources] --> provide a more detailed description of what the input dataset needs to look like. When I read [the written description](https://griml.readthedocs.io/en/stable/tutorials-processing.html#converting-raster-classifications-to-vectors) I understand that a person needs to bring a 2D raster with a binary classification of lake/not-lake. When I read the code below that description, though, it looks like a person needs a a 3-band tiff where band 1 is visible imagery ("rgb composite?") band 2 is SAR (any particular band?) and band 3 is a DEM.

Again you shouldn't feel compelled to support the second option, it is for sure more work. But I think you should pick one of the two directions and lean into that.

### Returning output data
In Python I think the most common pattern for functions that involve processing data structures usually that the input data structure is modified and returned to the user as a Python object, not saved to the file. I think it's great that your package saves the data structures to a file, but it would help the package feel more Pythonic if the dataframes were also returned to the user.

For some of your main functions it seems like this already is (or was at some point) the case:
- `convert()` and `merge_vectors()` looks like they already return dataframes
- `filter_vectors()` [has a return object](https://github.com/GEUS-Glaciology-and-Climate/GrIML/blob/848510271ed2482c27be311a1ec29aa1c838b6d6/src/griml/filter/filter_vectors.py#L24C7-L28C1) listed, but it doesn't look like the code returns anything

For the functions that do return dataframes I would encourage you to show that in the docs/notebooks (ex. `converted_dfs = convert([infile], proj, band_info, start, end)`). I know it isn't necessary for running the code, but for a first time user trying to explore the processing pipeline I think being able to look at the output of each step is helpful. For the core functions that don't have that I suggest returning the output dataframes. (`filter_vectors`, `add_metadata`)

### A few very minor notes
1. I love the background section (and the links section). It's a great resource. I'm noting here that there are a few images [on the background page](https://griml.readthedocs.io/en/stable/background.html) that aren't rendering right now.
2. The paper is really well written! One small note: Line 40 says "It is intended to continue using
the GrIML package for the production of future additions to...". To me "It is intended to continue using" is a bit awkward. Maybe something like "The authors intent to continue using the GRIML package for..." or "The GRIML package will continue to be used to produce future additions to the ...."
3. The second code cell in [the processing notebook](https://github.com/GEUS-Glaciology-and-Climate/GrIML/blob/main/tutorials/process_with_test_data.ipynb) has a line that I think should be commented `function can then be used to filter the vectors in the input file.` (the cell currently produces an error)

### An idea for the future
The notebooks that have been created of the docs pages are really helpful! Something to look into for the future is that many docs packages can render jupyter notebooks directly. [Here is the sphinx docs page](https://docs.readthedocs.com/platform/latest/guides/jupyter.html) describing how to use sphinx extensions for jupyter notebook rendering. You don't need to do this, but it would save you from needing to update your code in two places if it ever changes. In the meantime you could help users find your notebooks by linking to them at the beginning of each of the corresponding docs pages (Ex. A Jupyter notebook with this tutorial code is available here).

This was a lot of text, but I think the changes I've suggested are mostly quite small. This is a really nice package @PennyHow. I think it's really fantastic that you've decided to make it into this great community resource instead of keeping your own data and processing scripts hidden away on your own computer. Thanks for submitting it to JOSS!
