### 2025-06-11 2b4a6ae
Landing page description updated
