### 2025-06-12 6b5e748
JOSS badge added
