### 2025-06-12 1117f1a
[JOSS REVIEW] minor edits (#46)
* JOSS badge added

* wording change

* line comment removed
