+++ paper/paper.md
@@ -41,7 +41,7 @@ Initial rasterised binary classifications denoting water bodies can be inputted 
 
 There have been many different approaches to classifying ice-marginal lakes with remote sensing techniques [@How2021;@Chen2021;@Andreassen2022;@Rick2022;@Wieczorek2023;@Domgaard2024]. Packages exist for handling satellite and spatial data, such as GrIML's two key dependencies, Geopandas [@Geopandas2024] and Rasterio [@Gillies2019], as well as others such as SentinelHub [@Sentinelhub2024] and Google Earth Engine [@Gorelick2017]. Remote sensing object classification and post-processing routines are usually available in connection with scientific publications, however, few are available as open, deployable packages. The `GrIML` Python package addresses this gap, for the benefit of the future generation of ice-marginal lake inventories and for others in the scientific community to adapt and use themselves. 
 
-The 2016-2023 Greenland ice-marginal lake inventory series was produced with the `GrIML` Python package to address knowledge gaps regarding sea level contribution, and provide an overview of regions where lacustrine ice loss and processes are prominent [@How2025a;@How2025b]. The dataset consists of a series of annual inventories, mapping the presence and extent of ice-contact lakes across Greenland (\autoref{fig:inventory}). The annual inventory series spans the entirety of Greenland, including all terrestrial regions. Thus far, there are 8 annual inventories, covering 2016 to 2023, with one inventory for each year. It is intended to continue using the `GrIML` package for the production of future additions to the Greenland ice-marginal lake inventory series; in which case a standardised and version-controlled processing pipeline is essential [@How2025a].
+The 2016-2023 Greenland ice-marginal lake inventory series was produced with the `GrIML` Python package to address knowledge gaps regarding sea level contribution, and provide an overview of regions where lacustrine ice loss and processes are prominent [@How2025a;@How2025b]. The dataset consists of a series of annual inventories, mapping the presence and extent of ice-contact lakes across Greenland (\autoref{fig:inventory}). The annual inventory series spans the entirety of Greenland, including all terrestrial regions. Thus far, there are 8 annual inventories, covering 2016 to 2023, with one inventory for each year. The `GrIML` package will continue to be used to produce future additions to the Greenland ice-marginal lake inventory series; in which case a standardised and version-controlled processing pipeline is essential [@How2025a].
 
 ![An overview of the Greenland ice-marginal lake inventory series, which maps the number of lakes and their size between the years of 2016 and 2023 [@How2025a]. The map shows the location of each lake, with the symbol signifying area change over time, where green circles represent an increase in lake area, yellow squares represent a decrease in lake area, and white triangles represent an unchanged lake area. Crosshair points denote lakes where no area information is provided. The bar graphs show the change in number of lakes in each region. Base maps are provided by @Moon2023 and region outlines are as defined by @Mouginot2019. \label{fig:inventory}](https://github.com/GEUS-Glaciology-and-Climate/GrIML/blob/main/docs/figures/iml_dataset_overview.png?raw=true){ width=75% height=75% }
 
