### 2025-06-12 a5eb180
notebook reading requirements added
