+++ docs/index.rst
@@ -35,8 +35,8 @@ And include the following statement in the acknowledgements:
    
    installation
    background
-   tutorials-processing
-   tutorials-data
+   tutorials/process_with_test_data
+   tutorials/dataset_tutorials
    contributing
    acknowledgements
    modules
