### 2025-06-12 c26ed44
tutorials build routine added
