+++ .readthedocs.yaml
@@ -10,6 +10,10 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.8"
+  jobs:
+    pre-build:
+    - mkdir -p docs/tutorials
+    - cp tutorials/*.ipynb docs/tutorials/
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
