+++ .readthedocs.yaml
@@ -12,8 +12,8 @@ build:
     python: "3.8"
   jobs:
     pre-build:
-    - mkdir -p docs/tutorials
-    - cp tutorials/*.ipynb docs/tutorials/
+      - mkdir -p docs/tutorials
+      - cp tutorials/*.ipynb docs/tutorials/
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
