+++ .readthedocs.yaml
@@ -23,7 +23,7 @@ sphinx:
 # https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
   install:
-  - requirements: docs/requirements.txt
+    - requirements: docs/requirements.txt
 
 formats:
   - pdf
