+++ .readthedocs.yaml
@@ -12,7 +12,6 @@ build:
     python: "3.8"
   jobs:
     pre_build:
-      - mkdir -p docs/tutorials
       - cp tutorials/*.ipynb docs/tutorials/
 
 # Build documentation in the docs/ directory with Sphinx
