### 2025-06-16 f05df67
Jupyter notebook integration into readthedocs (#47)
* notebook integration extension added

* notebook reading requirements added

* symlinks to tutorials added

* index links updated

* Delete docs/tutorials directory

* tutorials build routine added

* indentation added

* pre_build typo

* filename typo

* indentation typo

* tutorials index mapping created

* tutorials index referred

* mkdir removed

* dataset description added

* Background docs moved

* index name changed

* tutorial contents migrated

* Delete docs/tutorials-data.md

* Delete docs/tutorials-processing.md
