+++ docs/background/index.rst
@@ -1,0 +1,8 @@
+Background
+==========
+
+.. toctree::
+   :maxdepth: 1
+   
+   Project description <background>
+   Dataset description <README-ESA-GRIML-IML-fv2>
