+++ docs/tutorials-processing.md
@@ -1,167 +1,0 @@
-# Processing tutorials
-
-The following tutorial steps make up the full workflow for post-processing of raster classifications to a compiled, production-ready inventory of ice marginal lakes.
-
-```{note}
-This tutorial is available as both a [python script](https://github.com/GEUS-Glaciology-and-Climate/GrIML/blob/main/test/process_with_test_data.py) and [jupyter notebook](https://mybinder.org/v2/gh/GEUS-Glaciology-and-Climate/GrIML/HEAD?urlpath=%2Fdoc%2Ftree%2Ftutorials%2Fdataset_tutorial.ipynb). We also provide a [Binder environment](https://mybinder.org/v2/gh/GEUS-Glaciology-and-Climate/GrIML/HEAD) within which you can run the notebook (with no need to set up the requirements yourself locally).
-```
-
-## Converting raster classifications to vectors
-
-GrIML's convert module is used to convert binary raster classifications of water bodies to vectors; where values of 1 in the raster classification denote water has been classified, and zero values denote no water. We need the `convert` function to perform this. 
-
-```python
-from griml.convert import convert
-```
-
-Next we need to define some input variables - the projection, band information and date range of the input raster. If you have followed the [GEE script for classifying lakes](https://github.com/GEUS-Glaciology-and-Climate/GrIML/blob/main/gee_scripts/lake_classification.js) available through the GrIMl repo, then each outputted raster band represents classifications using one of three approaches:
-
-1. Multi-spectral classification from Sentinel-2
-2. Backscatter threshold classification from Sentinel-1
-3. Sink detection from ArcticDEM
-
-And the default outputted projection is [Polar Stereographic](https://nsidc.org/data/user-resources/help-center/guide-nsidcs-polar-stereographic-projection). The start and end date should match the defined date range used for the raster classifications.
-
-```python
-# Define projection
-proj = 'EPSG:3413'
-
-# Define information from raster bands
-band_info = [{'b_number':1, 'method':'VIS', 'source':'S2'}, 
-             {'b_number':2, 'method':'SAR', 'source':'S1'},
-             {'b_number':3, 'method':'DEM', 'source':'ARCTICDEM'}] 
-
-# Define date range of classifications
-start='20170701' 
-end='20170831'
-```
-
-Next we need to define the location of our raster file. A test raster file is provided with GrIML, which can be located using the top level test directory in the repository.
-
-```python
-infile = 'test/test_data/test_north_greenland.tif'
-```
-
-And then the file can be converted from raster to vector classifications using the `convert` function and the input variables.
-
-```python
-vectors = convert([infile], proj, band_info, start, end) 
-```
-
-Note that a single raster file is wrapped as a list, as the `convert` function expects a list of rasters normally. We recommend using the `glob` module to generate a list of rasters from converting, if you wish to convert multiple classifications.
-
-```python
-# Generate list of all rasters
-import glob
-all_infiles = list(glob.glob("PATH/TO/RASTER_DIRECTORY/*.tif"))
-
-# Define output directory to write converted vectors to
-outdir = "PATH/TO/OUTPUT_DIRECTORY"
-
-# Convert all rasters 
-convert(all_infiles, proj, band_info, start, end, outdir) 
-```
-
-```{important}
-If an output directory is not provided then converted vectors will not be written to file.
-```
-
-
-## Filtering vector classifications
-
-Vector classifications can be filtered by size and proximity to an inputted mask. In our case, we remove water bodies below a size of 0.05 sq km and further than 1 km from the ice margin (as we want to retain larger ice marginal lakes).
-
[…]
