+++ .readthedocs.yaml
@@ -10,6 +10,9 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.8"
+  jobs:
+    pre_build:
+      - cp tutorials/*.ipynb docs/tutorials/
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
@@ -19,7 +22,7 @@ sphinx:
 # https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
 python:
   install:
-  - requirements: docs/requirements.txt
+    - requirements: docs/requirements.txt
 
 formats:
   - pdf
