### 2025-06-17 39833a8
Joss review/return output (#48)
* return variable added

* return object

* version bumped
