+++ docs/background/README-ESA-GRIML-IML-fv3.md
@@ -1,0 +1,100 @@
+# Greenland Ice-Marginal Lake Inventory annual time-series Edition 1
+
+Data available at [GEUS Dataverse](https://doi.org/10.22008/FK2/MBKW9N).
+
+Contact [pho@geus.dk](mailto:pho@geus.dk).
+
+## Table of contents
+
+- [Terms of use](#terms-of-use)
+- [Dataset Contents](#dataset-contents)
+	+ [Data format](#data-format)
+	+ [Metadata](#metadata)
+- [Acknowledgements](#acknowledgements)
+- [Relevant links](#relevant-links)
+
+## Terms of use
+
+If the data are presented or used to support results of any kind, please include references to the applicable publications:
+
+- *How, P. et al. (2025) "Greenland Ice-Marginal Lake Inventory annual time-series Edition 1". GEUS Dataverse. [https://doi.org/10.22008/FK2/MBKW9N](https://doi.org/10.22008/FK2/MBKW9N)*
+- *How, P. et al. (2021) "Greenland-wide inventory of ice marginal lakes using a multi-method approach". Sci. Rep. 11, 4481. [https://doi.org/10.1038/s41598-021-83509-1](https://doi.org/10.1038/s41598-021-83509-1)*
+- *How, P. et al. (In Review) "Greenland ice-marginal lake inventory series from 2016 to 2023". Earth Syst.Sci. Data Discuss. [https://doi.org/10.5194/essd-2025-18](https://doi.org/10.5194/essd-2025-18)*
+- *How, P. (In Review) "GrIML: A Python package for investigating Greenland's ice-marginal lakes under a changing climate". J. Open Source Software. [https://joss.theoj.org/papers/a2e10775df44b89f26b0ac9dbf8bc9e3](https://joss.theoj.org/papers/a2e10775df44b89f26b0ac9dbf8bc9e3)*
+
+And include the following statement in the acknowledgements:
+
+*"Ice-marginal lake data provided by the European Space Agency (ESA) and the Programme for Monitoring of the Greenland Ice Sheet (PROMICE) at the Geological Survey of Denmark and Greenland (GEUS) ([https://doi.org/10.22008/FK2/MBKW9N](https://doi.org/10.22008/FK2/MBKW9N))."*
+
+## Dataset contents
+
+This ice-marginal lake dataset is a series of annual inventories, mapping the extent and presence of lakes across Greenland that share a margin with the Greenland Ice Sheet and/or the surrounding ice caps and periphery glaciers. Specifically, the following files are included in this dataset:
+
+- *\<YYYYMMDD\>-ESA-GRIML-IML-\<version\>.gpkg*: Ice-marginal lake inventory for a specific year, provided as polygon vector features
+- *ALL-ESA-GRIML-IML-\<version\>.gpkg*: All classified ice-marginal lakes across the inventory series, which have been merged together and therefore one polygon vector features signifies one lake
+- *CURATED-ESA-GRIML-IML-\<version\>.gpkg*: All identified ice-marginal lakes across the inventory series, including manually-classified lakes, and provided as point vector features
+- *README-ESA-GRIML-IML-\<version\>.gpkg*: This dataset readme file
+
+The annual inventories provide a comprehensive record of all identified ice-marginal lakes, which have been detected using three independent remote sensing techniques:
+
+- DEM sink detection using the ArcticDEM (mosaic version 3)
+- SAR backscatter classification from Sentinel-1 imagery
+- Multi-spectral indices classification from Sentinel-2 imagery
+
[…]
