### 2025-07-04 6520ad7
Tutorial links updated
