### 2025-07-11 e9b982e
stable docs links updated
