### 2025-07-11 27a9fab
intermediary outputs >> results
