### 2025-07-11 f1227f6
Joss review/editor edits (#56)
* stable docs links updated

* binary classifications description added

* intermediary outputs >> results

* method availability changed
